+++ R/zzz.R
@@ -2,4 +2,6 @@
     ## requireNamespace("alabaster.base", quietly = TRUE)
     registerValidateObjectFunction("ms_backend_mz_r", validateMzBackendMzR)
     registerReadObjectFunction("ms_backend_mz_r", readMzBackendMzR)
+    registerValidateObjectFunction("spectra", validateSpectra)
+    registerReadObjectFunction("spectra", readSpectra)
 }